+++ Source/LowEntryExtendedStandardLibrary/Public/ClassesLatentAction/LowEntryLatentActionFloat.h
@@ -22,7 +22,7 @@ public:
 	UPROPERTY()
 	bool Finished = false;
 	UPROPERTY()
-	float Result = 0;
+	double Result = 0;
 
 	UPROPERTY()
 	int32 KeepAliveCount = 0;
@@ -32,7 +32,7 @@ public:
 	* Waits till the latent action is done.
 	*/
 	UFUNCTION(BlueprintCallable, Category = "Low Entry|Extended Standard Library|Latent Action|Float", meta = (Latent, WorldContext = "WorldContextObject", LatentInfo = "LatentInfo", DisplayName = "Wait Till Done", Keywords = "untill for end finished complete completion"))
-	void WaitTillDone(UObject* WorldContextObject, FLatentActionInfo LatentInfo, float& Result_);
+	void WaitTillDone(UObject* WorldContextObject, FLatentActionInfo LatentInfo, double& Result_);
 
 	void LatentActionDone();
 
@@ -40,7 +40,7 @@ public:
 	* Causes the latent action to be done.
 	*/
 	UFUNCTION(BlueprintCallable, Category = "Low Entry|Extended Standard Library|Latent Action|Float", Meta = (DisplayName = "Done", Keywords = "end finished complete completion"))
-	void Done(float Result_);
+	void Done(double Result_);
 
 	/**
 	* Returns true if the latent action is done.
@@ -52,5 +52,5 @@ public:
 	* Returns the result.
 	*/
 	UFUNCTION(BlueprintPure, Category = "Low Entry|Extended Standard Library|Latent Action|Float", Meta = (DisplayName = "Get Result", Keywords = "result return value"))
-	void GetResult(float& Result_);
+	void GetResult(double& Result_);
 };
